Fix potential race condition
Because data volume fields are saved on blur or on change it's possible
that a user could quickly submit two save requests, before the first has
been processed. This would result in the second request being denied
because it would have an out-of-date version number. The simplest and
most robust solution is to prevent any editing while the item is being
saved.

This commit also scraps use of the terribad typescript enums.